requirements.txt
flask-bcrypt==0.7.1
flask-restplus==0.13.0
flask-migrate==2.5.2
pyjwt==1.7.1
flask-script==2.0.6
flask_testing==0.7.1
gunicorn==20.0.4
werkzeug==0.16.0
